refactor(runtime): move vlm_components from exporters/ to runtime/
Pre-refactor janitor commit for the BaseVLA spine refactor (lift #1
Day 1 per features/03_export/basevla-spine_plan.md). The runtime
package imports vlm_components from the exporters package — a
cross-package coupling that the spine refactor will magnify if not
fixed first.

Per features/03_export/basevla-spine_research.md Lens 6 finding: the
coupling is exactly 1 file (vlm_components.py) imported by exactly
2 sites (runtime/vlm_orchestrator.py lines 235 + 435; tests/
test_vlm_prefix.py line 29). Moving the file to runtime/ where it
already conceptually lives makes the spine refactor's component
ownership story clean.

Changes:
- git mv src/reflex/exporters/vlm_components.py
        src/reflex/runtime/vlm_components.py
- src/reflex/runtime/vlm_orchestrator.py:235 — import path updated
- src/reflex/runtime/vlm_orchestrator.py:435 — import path updated
- tests/test_vlm_prefix.py:29 — import path updated

Behavior-preserving. No public-API change (these symbols were never
in src/reflex/exporters/__init__.py's surface). Zero new dependencies.
